Rebuild without sockets, less callbacks and session middleware.
The state is stored in a session, which is identified by a signed
cookie. No data is manually stored in the DB anymore except the list of
public chats.
The browser polls for status updates every 5 seconds.
The server asks DC directly, if the wanted group has other members than
the bot. Thus no callbacks are necessary to inform the app, or the
browser of the successful join.